+++ packages/atproto/lib/src/entities/invite_code.freezed.dart
@@ -1,0 +1,151 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'invite_code.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+InviteCode _$InviteCodeFromJson(Map<String, dynamic> json) {
+  return _InviteCode.fromJson(json);
+}
+
+/// @nodoc
+mixin _$InviteCode {
+  String get code => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $InviteCodeCopyWith<InviteCode> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $InviteCodeCopyWith<$Res> {
+  factory $InviteCodeCopyWith(
+          InviteCode value, $Res Function(InviteCode) then) =
+      _$InviteCodeCopyWithImpl<$Res, InviteCode>;
+  @useResult
+  $Res call({String code});
+}
+
+/// @nodoc
+class _$InviteCodeCopyWithImpl<$Res, $Val extends InviteCode>
+    implements $InviteCodeCopyWith<$Res> {
+  _$InviteCodeCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? code = null,
+  }) {
+    return _then(_value.copyWith(
+      code: null == code
+          ? _value.code
+          : code // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$_InviteCodeCopyWith<$Res>
+    implements $InviteCodeCopyWith<$Res> {
+  factory _$$_InviteCodeCopyWith(
+          _$_InviteCode value, $Res Function(_$_InviteCode) then) =
+      __$$_InviteCodeCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String code});
+}
+
+/// @nodoc
+class __$$_InviteCodeCopyWithImpl<$Res>
+    extends _$InviteCodeCopyWithImpl<$Res, _$_InviteCode>
+    implements _$$_InviteCodeCopyWith<$Res> {
+  __$$_InviteCodeCopyWithImpl(
+      _$_InviteCode _value, $Res Function(_$_InviteCode) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? code = null,
+  }) {
+    return _then(_$_InviteCode(
+      code: null == code
+          ? _value.code
+          : code // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_InviteCode implements _InviteCode {
+  const _$_InviteCode({required this.code});
+
+  factory _$_InviteCode.fromJson(Map<String, dynamic> json) =>
+      _$$_InviteCodeFromJson(json);
+
+  @override
+  final String code;
+
+  @override
+  String toString() {
+    return 'InviteCode(code: $code)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$_InviteCode &&
+            (identical(other.code, code) || other.code == code));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hash(runtimeType, code);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$_InviteCodeCopyWith<_$_InviteCode> get copyWith =>
+      __$$_InviteCodeCopyWithImpl<_$_InviteCode>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_InviteCodeToJson(
+      this,
+    );
+  }
+}
+
+abstract class _InviteCode implements InviteCode {
+  const factory _InviteCode({required final String code}) = _$_InviteCode;
+
+  factory _InviteCode.fromJson(Map<String, dynamic> json) =
+      _$_InviteCode.fromJson;
+
+  @override
+  String get code;
+  @override
+  @JsonKey(ignore: true)
+  _$$_InviteCodeCopyWith<_$_InviteCode> get copyWith =>
+      throw _privateConstructorUsedError;
+}