+++ packages/atproto/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.6
 
 - Supported `com.atproto.server.createAccount`. ([#256](https://github.com/myConsciousness/atproto.dart/issues/256))
+- Supported `com.atproto.server.createInviteCode`. ([#257](https://github.com/myConsciousness/atproto.dart/issues/257))
 
 ## v0.2.5